natflow: fix speed rate calculation accuracy

+++ natflow_user.c
@@ -904,7 +904,7 @@ static unsigned int natflow_user_pre_hook(void *priv,
 						unsigned int rx_speed_jiffies = atomic_read(&fud->rx_speed_jiffies);
 						unsigned int tx_speed_jiffies = atomic_read(&fud->tx_speed_jiffies);
 						int x = (jiffies/HZ/2) % 4;
-						unsigned long diff = uintmindiff(jiffies, rx_speed_jiffies);
+						unsigned int diff = uintmindiff(jiffies, rx_speed_jiffies);
 						if (diff > HZ * 2 * 4) {
 							user_i->rx_speed_packets = 0;
 							user_i->rx_speed_bytes = 0;
@@ -1324,9 +1324,9 @@ static unsigned int natflow_user_post_hook(void *priv,
 		if (user->tuplehash[IP_CT_DIR_ORIGINAL].tuple.src.u3.ip != ct->tuplehash[CTINFO2DIR(ctinfo)].tuple.src.u3.ip) {
 			/* download */
 			unsigned int rx_speed_jiffies = atomic_xchg(&fud->rx_speed_jiffies, jiffies);
-			int i = (jiffies/HZ/2) % 4;
-			int j = (rx_speed_jiffies/HZ/2) % 4;
-			unsigned long diff_jiffies = uintmindiff(jiffies, rx_speed_jiffies);
+			unsigned int i = ((unsigned int)jiffies/HZ/2) % 4;
+			unsigned int j = (rx_speed_jiffies/HZ/2) % 4;
+			unsigned int diff_jiffies = uintmindiff(jiffies, rx_speed_jiffies);
 			if (diff_jiffies >= HZ * 2 * 4) {
 				for(j = 0; j < 4; j++) {
 					atomic_set(&fud->rx_speed_bytes[j], 0);
@@ -1346,9 +1346,9 @@ static unsigned int natflow_user_post_hook(void *priv,
 		} else {
 			/* upload */
 			unsigned int tx_speed_jiffies = atomic_xchg(&fud->tx_speed_jiffies, jiffies);
-			int i = (jiffies/HZ/2) % 4;
-			int j = (tx_speed_jiffies/HZ/2) % 4;
-			unsigned long diff_jiffies = uintmindiff(jiffies, tx_speed_jiffies);
+			unsigned int i = ((unsigned int)jiffies/HZ/2) % 4;
+			unsigned int j = (tx_speed_jiffies/HZ/2) % 4;
+			unsigned int diff_jiffies = uintmindiff(jiffies, tx_speed_jiffies);
 			if (diff_jiffies >= HZ * 2 * 4) {
 				for(j = 0; j < 4; j++) {
 					atomic_set(&fud->tx_speed_bytes[j], 0);
@@ -2084,7 +2084,7 @@ static ssize_t userinfo_read(struct file *file, char __user *buf,
 						unsigned int rx_speed_jiffies = atomic_read(&fud->rx_speed_jiffies);
 						unsigned int tx_speed_jiffies = atomic_read(&fud->tx_speed_jiffies);
 						int x = (jiffies/HZ/2) % 4;
-						unsigned long diff = uintmindiff(jiffies, rx_speed_jiffies);
+						unsigned int diff = uintmindiff(jiffies, rx_speed_jiffies);
 						if (diff > HZ * 2 * 4) {
 							user_i->rx_speed_packets = 0;
 							user_i->rx_speed_bytes = 0;
